requirements.txt
astroid==2.9.0
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cycler==0.11.0
fonttools==4.28.2
idna==3.3
isort==5.10.1
joblib==1.1.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
matplotlib==3.5.0
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pathspec==0.9.0
Pillow==8.4.0
pkg_resources==0.0.0
platformdirs==2.4.0
praw==7.5.0
prawcore==2.3.0
pylint==2.12.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
update-checker==0.18.0
urllib3==1.26.7
websocket-client==1.2.1
wrapt==1.13.3
